splitted tests and annotation, added line to contributing file, showing both annoations
